Reshuffle some sorting code from processor.go
