+++ sort.go
@@ -1,5 +1,7 @@
 package aggro
 
+import "sort"
+
 // Sortable provides an interface that various sorters can implement to compare
 // two result buckets. This enables access to both the value and any metrics
 // that the query may have contained.
@@ -30,4 +32,53 @@ type AlphabeticalSortable bool
 // Less implements Sortable by comparing the values field of each result.
 func (sortable *AlphabeticalSortable) Less(a, b *ResultBucket) bool {
 	return a.Value < b.Value == bool(*sortable)
+}
+
+// bucketSorter is an implementation of the sort.Sort interface that is capable
+// or sorting the supplied slice of results with the supplied Sortable.
+type bucketSorter struct {
+	results  []*ResultBucket
+	sortable Sortable
+}
+
+// Len implements the sort.Sort interface Len method.
+func (sorter *bucketSorter) Len() int {
+	return len(sorter.results)
+}
+
+// Swap implements the sort.Sort interface Swap method.
+func (sorter *bucketSorter) Swap(i, j int) {
+	sorter.results[i], sorter.results[j] = sorter.results[j], sorter.results[i]
+}
+
+// Less implement the sort.Sort less method by calling the Sortable that the
+// sorter has been supplied, with the two appropriate results.
+func (sorter *bucketSorter) Less(i, j int) bool {
+	return sorter.sortable.Less(sorter.results[i], sorter.results[j])
+}
+
+// sortMap takes a map of results and returns it as a sorted slice.
+func sortMap(bucket *Bucket, results map[string]*ResultBucket) []*ResultBucket {
+	resultSlice := []*ResultBucket{}
+	for _, result := range results {
+		resultSlice = append(resultSlice, result)
+	}
+	return sortSlice(bucket, resultSlice)
+}
+
+// sortSlice takes a slice of results and sorts it via a bucketSorter instance.
+func sortSlice(bucket *Bucket, results []*ResultBucket) []*ResultBucket {
+	sorter := &bucketSorter{
+		results:  results,
+		sortable: sortableForOptions(bucket.Sort),
+	}
+	if sorter.sortable != nil {
+		sort.Sort(sorter)
+	}
+	for _, result := range sorter.results {
+		if bucket.Bucket != nil {
+			result.Buckets = sortMap(bucket.Bucket, result.bucketLookup)
+		}
+	}
+	return sorter.results
 }
